requirements.txt
# RAG评估系统依赖包
# 核心框架
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# 文本处理和相似度计算
nltk>=3.7
jieba>=0.42.1

# LLM和Embedding
ragas==0.3.2
langchain>=0.1.0,<0.3.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0

# 数据库相关
pymysql>=1.0.2
SQLAlchemy>=1.4.0
alembic>=1.8.0

# 网络请求
requests>=2.25.0
httpx>=0.24.0

# 配置管理
python-dotenv>=0.19.0

# 表格显示
tabulate>=0.9.0

# 文件处理
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# 异步支持
asyncio-mqtt>=0.11.0

# 其他工具
pathlib2>=2.3.0
typing-extensions>=4.0.0
